+++ transektcount/src/main/java/com/wmstein/transektcount/EditMetaLActivity.java
@@ -5,9 +5,9 @@ import android.app.DatePickerDialog;
 import android.app.TimePickerDialog;
 import android.content.SharedPreferences;
 import android.graphics.Bitmap;
+import android.graphics.Typeface;
 import android.graphics.drawable.BitmapDrawable;
 import android.os.Bundle;
-import android.text.Html;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -36,7 +36,7 @@ import androidx.appcompat.app.AppCompatActivity;
 /***************************************************************
  * EditMetaActivity collects meta info for a transect inspection
  * Created by wmstein on 2016-03-31,
- * last edited on 2023-05-09
+ * last edited on 2023-06-11
  */
 public class EditMetaLActivity extends AppCompatActivity implements SharedPreferences.OnSharedPreferenceChangeListener
 {
@@ -285,27 +285,30 @@ public class EditMetaLActivity extends AppCompatActivity implements SharedPrefer
         meta.tempe = etw.getWidgetTemp();
         if (meta.tempe > 50 || meta.tempe < 0)
         {
-            Snackbar sB = Snackbar.make(etw, Html.fromHtml("<font color=\"#ff0000\"><b>" +  getString(R.string.valTemp) + "</font></b>"), Snackbar.LENGTH_LONG);
+            Snackbar sB = Snackbar.make(etw, getString(R.string.valTemp), Snackbar.LENGTH_LONG);
             TextView tv = sB.getView().findViewById(R.id.snackbar_text);
             tv.setTextAlignment(View.TEXT_ALIGNMENT_TEXT_START);
+            tv.setTypeface(tv.getTypeface(), Typeface.BOLD);
             sB.show();
             return false;
         }
         meta.wind = etw.getWidgetWind();
         if (meta.wind > 4 || meta.wind < 0)
         {
-            Snackbar sB = Snackbar.make(etw, Html.fromHtml("<font color=\"#ff0000\"><b>" +  getString(R.string.valWind) + "</font></b>"), Snackbar.LENGTH_LONG);
+            Snackbar sB = Snackbar.make(etw, getString(R.string.valWind), Snackbar.LENGTH_LONG);
             TextView tv = sB.getView().findViewById(R.id.snackbar_text);
-            tv.setTextAlignment(View.TEXT_ALIGNMENT_CENTER);
+            tv.setTextAlignment(View.TEXT_ALIGNMENT_TEXT_START);
+            tv.setTypeface(tv.getTypeface(), Typeface.BOLD);
             sB.show();
             return false;
         }
         meta.clouds = etw.getWidgetClouds();
         if (meta.clouds > 100 || meta.clouds < 0)
         {
-            Snackbar sB = Snackbar.make(etw, Html.fromHtml("<font color=\"#ff0000\"><b>" +  getString(R.string.valClouds) + "</font></b>"), Snackbar.LENGTH_LONG);
+            Snackbar sB = Snackbar.make(etw, getString(R.string.valClouds), Snackbar.LENGTH_LONG);
             TextView tv = sB.getView().findViewById(R.id.snackbar_text);
-            tv.setTextAlignment(View.TEXT_ALIGNMENT_TEXT_END);
+            tv.setTextAlignment(View.TEXT_ALIGNMENT_TEXT_START);
+            tv.setTypeface(tv.getTypeface(), Typeface.BOLD);
             sB.show();
             return false;
         }