+++ transektcount/src/main/java/com/wmstein/transektcount/MyDebug.kt
@@ -1,0 +1,11 @@
+package com.wmstein.transektcount
+
+/**
+ * Global constant to control Logcat logging,
+ * should only temporarily set to true for debugging,
+ * set always to false for release versions.
+ */
+object MyDebug {
+    const val LOG = false
+//    const val LOG = true
+}