+++ transektcount/src/main/java/com/wmstein/transektcount/NewSectionLActivity.java
@@ -12,7 +12,6 @@ import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.ViewGroup;
 import android.view.WindowManager;
 import android.widget.EditText;
 import android.widget.ScrollView;
@@ -50,7 +49,7 @@ public class NewSectionLActivity extends AppCompatActivity implements SharedPref
     Section section;
     Section newSection;
     
-    ViewGroup layout;
+//    ViewGroup layout;
     EditText newsectName;
     private SectionDataSource sectionDataSource;
     List<Section> sections = new ArrayList<>();
@@ -118,14 +117,14 @@ public class NewSectionLActivity extends AppCompatActivity implements SharedPref
         int id = item.getItemId();
         if (id == R.id.menuSaveExit)
         {
-            saveSection(layout);
+            saveSection();
         }
         return super.onOptionsItemSelected(item);
     }
 
     // Save section with plausi-check for empty or duplicate section name
     @SuppressLint("ApplySharedPref")
-    public void saveSection(View view)
+    public void saveSection()
     {
         // first, edit the section name
         String sect_name = newsectName.getText().toString();
@@ -236,7 +235,7 @@ public class NewSectionLActivity extends AppCompatActivity implements SharedPref
     {
         View view = findViewById(R.id.newsectScreen);
         Snackbar sB = Snackbar.make(view, str, Snackbar.LENGTH_LONG);
-        sB.setActionTextColor(Color.RED);
+        sB.setTextColor(Color.RED);
         TextView tv = sB.getView().findViewById(R.id.snackbar_text);
         tv.setTextAlignment(View.TEXT_ALIGNMENT_CENTER);
         tv.setTypeface(tv.getTypeface(), Typeface.BOLD);